requirements.txt
mlx>=0.3.0
mlx-lm>=0.0.10
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
transformers>=4.36.0
datasets>=2.16.0
sentence-transformers>=2.2.2
langchain>=0.1.0
ollama>=0.1.0
fastapi>=0.95.2
uvicorn>=0.22.0
matplotlib>=3.7.0
